requirements.txt
torch
transformers
datasets
accelerate
deepspeed
pillow
numpy
tqdm
pyyaml
liger-kernel
qwen-vl-utils
